Refactored data structures
