+++ alkos/kernel/test/template_lib_test.cpp
@@ -1,7 +1,9 @@
+#include <extensions/data_structures/data_structures.hpp>
 #include <extensions/template_lib.hpp>
 #include <test_module/test.hpp>
 
 using namespace template_lib;
+using namespace data_structures;
 
 class TemplateLibTest : public TestGroupBase
 {
