+++ alkos/libc/include/extensions/data_structures/data_structures.hpp
@@ -1,4 +1,7 @@
 #ifndef ALKOS_LIBC_INCLUDE_EXTENSIONS_DATA_STRUCTURES_DATA_STRUCTURES_HPP_
 #define ALKOS_LIBC_INCLUDE_EXTENSIONS_DATA_STRUCTURES_DATA_STRUCTURES_HPP_
 
+#include <extensions/data_structures/array_structures.hpp>
+#include <extensions/data_structures/events.hpp>
+
 #endif  // ALKOS_LIBC_INCLUDE_EXTENSIONS_DATA_STRUCTURES_DATA_STRUCTURES_HPP_
